requirements.txt
ruff==0.14.10
requests==2.32.5
python-dotenv==1.2.1
pyinstaller==6.17.0
pyodbc==5.3.0
pandas==2.3.3
xlsxwriter==3.2.9
